Lots of work on autocomplete and more refactoring code.
Autocomplete is nearly done, there is a new script to handle
autocomplete as a jquery plugin (currently uses bootstrap). The main
query.js code has been refactored to be inside an anonymous function to
permit use of "use strict" on all the code. Minimal functionality
changes other than a schema update.

Future: The actual retrieval code will need to be rewritten for the
new paging. This will actually make the code more efficient and easier
to read.

Change-Id: I1abf71469f56390ae7a3db9c4fb8889a756a40e0