Migration to new build pattern and new retrieval code
This also includes lots of minor fixes to just about everything.
The migration only includes moving files for now but will include new
build steps in the future.

For the time being please use /catalog-dev, it is unfiltered original
code. The build script will be building into /catalog in the future.

Change-Id: Ib49515233aa5e7ebfdf1385b763f986341d82f4a